+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **v0.2.1:**
 - Bug fix: spec update (PR #7): blocklist filtering in uppercase-only alphabet [[PR #7](https://github.com/sqids/sqids-spec/pull/7)]
+- Lower uniques test from 1_000_000 to 10_000
 
 **v0.2.0:**
 - Bug fix: test for decoding an invalid ID with a [repeating reserved character](https://github.com/sqids/sqids-spec/commit/f52b57836b0463097018f984f853b284e50a5ce4)